refactor: factor out SearchSelectUserCommand into its own component for reusability
